move components, lib, globals.css inside app
