+++ app/lib/searchUtils.ts
